Add directories + fix imports
