+++ lib/app/core/utils/global_bindings.dart
@@ -1,0 +1,18 @@
+import 'package:get/get.dart';
+import 'package:tutorial_inicial/app/data/dao/category_dao.dart';
+import 'package:tutorial_inicial/app/data/repositories/category_list_repository.dart';
+import 'package:tutorial_inicial/app/modules/category_list/unit_dao.dart';
+import 'package:tutorial_inicial/app/modules/unit_converter/unit_converter_controller.dart';
+
+
+class GlobalBindings extends Bindings {
+  @override
+  void dependencies() {
+    Get.put<CategoryDao>(CategoryDao(), permanent: true);
+    Get.put<UnitDao>(UnitDao(), permanent: true);
+
+    Get.put<CategoryListRepository>(CategoryListRepository(), permanent: true);
+
+    Get.lazyPut<UnitConverterController>(() => UnitConverterController());
+  }
+}
