+++ lib/app/data/repositories/category_list_repository.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
-import 'package:tutorial_inicial/app/category_list/category_dao.dart';
-import 'package:tutorial_inicial/app/category_list/unit_dao.dart';
+import 'package:tutorial_inicial/app/data/dao/category_dao.dart';
+import 'package:tutorial_inicial/app/data/dao/unit_dao.dart';
 import 'package:tutorial_inicial/app/data/model/unit.dart';
 
 class CategoryListRepository {
