+++ lib/app/modules/category_list/category_list.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:tutorial_inicial/app/category_list/category_list_controller.dart';
-import 'package:tutorial_inicial/app/category_list/category_tile.dart';
 import 'package:tutorial_inicial/app/data/model/category.dart';
-import 'package:tutorial_inicial/app/info/info.dart';
+import 'package:tutorial_inicial/app/modules/category_list/category_list_controller.dart';
+import 'package:tutorial_inicial/app/modules/category_list/local_widgets/category_tile.dart';
+import 'package:tutorial_inicial/app/pages/info/info.dart';
+
 
 class CategoryList extends GetView<CategoryListController> {
   const CategoryList({Key? key}) : super(key: key);
