+++ lib/app/modules/category_list/local_widgets/category_tile.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:tutorial_inicial/app/data/model/category.dart';
-import 'package:tutorial_inicial/app/unit_converter/unit_converter.dart';
+import 'package:tutorial_inicial/app/modules/unit_converter/unit_converter.dart';
 
 class CategoryTile extends StatelessWidget {
   final Category category;
