+++ lib/app/pages/welcome/welcome.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:tutorial_inicial/app/category_list/category_list.dart';
+import 'package:tutorial_inicial/app/modules/category_list/category_list.dart';
 
 class Welcome extends StatelessWidget {
   const Welcome({Key? key}) : super(key: key);
