+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:tutorial_inicial/app/utils/global_bindings.dart';
+import 'package:tutorial_inicial/app/core/utils/global_bindings.dart';
 import 'app/routes/pages.dart';
 import 'app/routes/routes.dart';
